requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2
pandas
pyarrow
scikit-learn
lightgbm
numpy
tqdm
streamlit
firebase-admin
schedule

